+++ PRTelegramBot/Workflow/Interfaces/IWorkflowCondition.cs
@@ -1,0 +1,10 @@
+﻿using PRTelegramBot.Interfaces;
+
+namespace PRTelegramBot.Workflow.Interfaces
+{
+    public interface IWorkflowCondition : IWorkflowNode
+    {
+        Task<bool> CheckConditionAsync(IBotContext context);
+    }
+}
+