+++ PRTelegramBot/Workflow/Interfaces/IWorkflowNode.cs
@@ -1,0 +1,6 @@
+﻿namespace PRTelegramBot.Workflow.Interfaces
+{
+    public interface IWorkflowNode
+    {
+    }
+}