+++ PRTelegramBot/Workflow/Interfaces/Tasks/IWorkflowManulTask.cs
@@ -1,0 +1,6 @@
+﻿namespace PRTelegramBot.Workflow.Interfaces.Tasks
+{
+    public interface IWorkflowManulTask
+    {
+    }
+}