+++ PRTelegramBot/Workflow/TelegramStateBuilder.cs
@@ -1,0 +1,10 @@
+﻿namespace PRTelegramBot.Workflow
+{
+    internal class TelegramStateBuilder
+    {
+        public TelegramStateManager Build()
+        {
+            return new TelegramStateManager();
+        }
+    }
+}