+++ PRTelegramBot/Workflow/TelegramStateManager.cs
@@ -1,0 +1,6 @@
+﻿namespace PRTelegramBot.Workflow
+{
+    public class TelegramStateManager
+    {
+    }
+}